requirements.txt
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.4.1
dgl-cu111
googledrivedownloader==0.4
idna==3.10
importlib-metadata==4.13.0
isodate==0.6.1
Jinja2==3.1.5
joblib==1.3.2
littleutils==0.2.2
MarkupSafe==2.1.5
networkx==2.6.3
numpy==1.21.6
ogb==1.3.6
outdated==0.2.2
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
protobuf==4.24.4
psutil==7.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.1
rdflib==6.3.2
requests==2.31.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.17.0
tensorboardX==2.6.2.2
threadpoolctl==3.1.0
torch==1.9.0+cu111
torch-cluster
torch-geometric==2.0.3
torch-scatter
torch-sparse
torch-spline-conv
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.67.1
typing_extensions==4.7.1
urllib3==2.0.7
yacs==0.1.8
zipp==3.15.0
